+++ tests/specs/basic.spec.js
@@ -6,7 +6,7 @@ describe('Drag drop', () => {
 
     cy.findByTestId('left').find('.item').assertList(['1', '2', '3', '4', '5', '6'])
 
-    cy.findByTestId('item-1').drag('[data-testid="right"]')
+    cy.findByTestId('item-1').drag('[data-testid="right"]', { target: { position: 'left' } })
 
     cy.findByTestId('right').find('.item').assertList(['1'])