+++ tests/specs/move.spec.js
@@ -10,7 +10,7 @@ describe('Drag drop', () => {
       expect(left).to.be.equal(100)
     })
 
-    cy.findByTestId('draggable').move({ x: 100, y: 100, position: 'center' })
+    cy.findByTestId('draggable').move({ deltaX: 100, deltaY: 100 })
 
     cy.findByTestId('draggable').should('have.css', 'transform', 'matrix(1, 0, 0, 1, 100, 100)')
     cy.findByTestId('draggable').should((target) => {
@@ -31,7 +31,7 @@ describe('Drag drop', () => {
       expect(left).to.be.equal(100)
     })
 
-    cy.findByTestId('draggable').move({ x: 100, y: 100, position: 'center' })
+    cy.findByTestId('draggable').move({ deltaX: 100, deltaY: 100 })
 
     cy.findByTestId('draggable').should('have.css', 'transform', 'matrix(1, 0, 0, 1, 100, 100)')
     cy.findByTestId('draggable').should((target) => {